+++ relay-profiling/src/lib.rs
@@ -122,7 +122,7 @@ pub fn parse_metadata(payload: &[u8], project_id: ProjectId) -> Result<ProfileId
                 {
                     Ok(profile) => profile,
                     Err(err) => {
-                        relay_log::warn!(
+                        relay_log::debug!(
                             error = &err as &dyn Error,
                             from = "metadata",
                             platform = "android",
@@ -188,7 +188,7 @@ pub fn expand_profile(payload: &[u8], event: &Event) -> Result<(ProfileId, Vec<u
                 Err(ProfileError::InvalidJson(err))
             }
             _ => {
-                relay_log::warn!(
+                relay_log::debug!(
                     error = &err as &dyn Error,
                     from = "parsing",
                     platform = profile.platform,